# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API BlogAtr",
        "description": "Lorem Ipsum",
        "contact": {
            "email": "arifingdr@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "User",
                "description": "Mengambil Data Semua User",
                "operationId": "User All",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": [
                                        {
                                            "id": 6,
                                            "name": "azriel",
                                            "email": "azriel@lenna.ai",
                                            "book": [],
                                            "created_at": "2024-02-20T04:23:35.000000Z",
                                            "updated_at": "2024-02-20T04:23:35.000000Z"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Book"
                ],
                "summary": "Book",
                "description": "Membuat Data Book",
                "operationId": "Update Book",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter ID Book",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "user_id": "1",
                                    "name": "arifin"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": {
                                        "name": "azriel",
                                        "user_id": "6"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User",
                "description": "Membuat Data User",
                "operationId": "Create User",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Jessica Smith",
                                    "email": "arifin@gmail.com",
                                    "password": "arifin12",
                                    "password_confirmation": "arifin12"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": {
                                        "id": 6,
                                        "name": "azriel",
                                        "email": "azriel@lenna.ai",
                                        "book": [],
                                        "created_at": "2024-02-20T04:23:35.000000Z",
                                        "updated_at": "2024-02-20T04:23:35.000000Z"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "User",
                "description": "Mengambil Data Dari ID User",
                "operationId": "User By Id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter ID User",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": {
                                        "id": 6,
                                        "name": "azriel",
                                        "email": "azriel@lenna.ai",
                                        "book": [],
                                        "created_at": "2024-02-20T04:23:35.000000Z",
                                        "updated_at": "2024-02-20T04:23:35.000000Z"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "User",
                "description": "Membuat Data User",
                "operationId": "Remove User",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter ID User",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": []
                                }
                            }
                        }
                    }
                }
            }
        },
        "/createBook": {
            "post": {
                "tags": [
                    "Book"
                ],
                "summary": "Book",
                "description": "Membuat Data Book",
                "operationId": "Create Book",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "user_id": "1",
                                    "name": "arifin"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": {
                                        "name": "azriel",
                                        "user_id": "6"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}